+++ src/PluginInfo.cs
@@ -4,6 +4,6 @@ namespace PitFireTeamFikaFix
     {
         public const string GUID = "com.dematch.pitfireteam-fika";
         public const string NAME = "Pit Fire Team Fika Fix (headless_host)";
-        public const string VERSION = "0.2.1";
+        public const string VERSION = "0.2.2";
     }
 }